Added Texture class and bitmap loading functionality.
ImageLoader.hpp and .cpp contain the bitmap loading function, and these should contain any other image loading functionality. This is to distinguish between the particulars of a given file format and the RGB or RGBA array that OpenGL expects. For now, only 24 bpp, uncompressed bitmaps can be loaded (this is a very common format, though). Note: texture loading should be done after OpenGL has been initialized and a window has been created with GLUT. Just because.