Added Window class and Event handling.
Window class abstracts a bunch of GLUT-related callback stuff away from main, and makes it so you can initialize a window by instantiating the class. All rendering that main wants to do should be defined in the renderScene() private static member function of the Window class. It is intentionally left undefined in Window.cpp so that it can be defined as needed to suit rendering needs. Also, the Windowing system abstracts away the keyboard/mouse event callbacks, so that now you can use event polling instead, which also takes GLUT-related gunk out of the main.cpp file. Note that a call to glutPostRedisplay() should be made in whatever event handling function you implement, as without it the window will never update. See the included main.cpp demo for a simple example.