Made key events more explicit with what key is pressed.
Note that if an ASCII key was pressed, then Event.keyData.isAscii will be true and you can simply extract the key's ASCII value from Event.keyData.keyCode. If it's not an ASCII key that was pressed (say, F1 or an arrow key), then Event.keyData.isAscii will be false, and while you can still get a value from Event.keyData.keyCode, it will contain a value which will be bogus to ASCII interpretation.